+++ opendmi/src/format/yaml/helpers.c
@@ -5,31 +5,65 @@
 // SPDX-License-Identifier: BSD-3-Clause
 //
 #include <string.h>
+#include <ctype.h>
 #include <assert.h>
+
+#include <opendmi/error.h>
+#include <opendmi/utils.h>
 
 #include <opendmi/format/yaml/helpers.h>
 
-bool dmi_yaml_label(dmi_yaml_session_t *session, const char *value)
+bool dmi_yaml_emit(dmi_yaml_session_t *session, yaml_event_t *event)
 {
-    return dmi_yaml_scalar(session, value, YAML_PLAIN_SCALAR_STYLE);
+    assert(session != nullptr);
+    assert(event != nullptr);
+
+    bool result = yaml_emitter_emit(session->emitter, event);
+    if (not result) {
+        dmi_error_raise_ex(session->context, DMI_ERROR_INTERNAL,
+                           "Unable to emit YAML event: %s", session->emitter->problem);
+    }
+
+    return result;
 }
 
-bool dmi_yaml_scalar(dmi_yaml_session_t *session, const char *value, yaml_scalar_style_t style)
+bool dmi_yaml_label(dmi_yaml_session_t *session, const char *value)
+{
+    return dmi_yaml_scalar(session, value, nullptr, YAML_PLAIN_SCALAR_STYLE);
+}
+
+bool dmi_yaml_scalar(
+        dmi_yaml_session_t  *session,
+        const char          *value,
+        const char          *tag,
+        yaml_scalar_style_t  style)
 {
     assert(session != nullptr);
 
     bool success = false;
     yaml_event_t event = {};
+    char *escaped = nullptr;
 
     do {
-        if (not yaml_scalar_event_initialize(
-                    &event, nullptr, nullptr, (const yaml_char_t *)value, strlen(value), true, false, style))
+        bool result = yaml_scalar_event_initialize(&event, nullptr,
+                                                   (yaml_char_t *)tag,
+                                                   (yaml_char_t *)value,
+                                                   strlen(value),
+                                                   true, false, style);
+
+        if (not result) {
+            dmi_error_raise_ex(session->context, DMI_ERROR_INTERNAL,
+                               "Unable to initialize scalar event: %s",
+                               session->emitter->problem);
             break;
-        if (not yaml_emitter_emit(session->emitter, &event))
+        }
+        if (not dmi_yaml_emit(session, &event))
             break;
 
         success = true;
     } while (false);
+
+    dmi_free(escaped);
 
     return success;
 }
@@ -42,9 +76,17 @@ bool dmi_yaml_sequence_start(dmi_yaml_session_t *session, yaml_sequence_style_t 
     yaml_event_t event = {};
 
     do {
-        if (not yaml_sequence_start_event_initialize(&event, nullptr, nullptr, true, style))
+        bool result = yaml_sequence_start_event_initialize(&event, nullptr,
+                                                           (yaml_char_t *)YAML_SEQ_TAG,
+                                                           true, style);
+
+        if (not result) {
+            dmi_error_raise_ex(session->context, DMI_ERROR_INTERNAL,
+                               "Unable to initialize sequence start event: %s",
+                               session->emitter->problem);
             break;
-        if (not yaml_emitter_emit(session->emitter, &event))
+        }
+        if (not dmi_yaml_emit(session, &event))
             break;
 
         success = true;
@@ -63,7 +105,7 @@ bool dmi_yaml_sequence_end(dmi_yaml_session_t *session)
     do {
         if (not yaml_sequence_end_event_initialize(&event))
             break;
-        if (not yaml_emitter_emit(session->emitter, &event))
+        if (not dmi_yaml_emit(session, &event))
             break;
 
         success = true;
@@ -80,9 +122,9 @@ bool dmi_yaml_mapping_start(dmi_yaml_session_t *session, yaml_mapping_style_t st
     yaml_event_t event = {};
 
     do {
-        if (not yaml_mapping_start_event_initialize(&event, nullptr, nullptr, true, style))
+        if (not yaml_mapping_start_event_initialize(&event, nullptr, (yaml_char_t *)YAML_MAP_TAG, true, style))
             break;
-        if (not yaml_emitter_emit(session->emitter, &event))
+        if (not dmi_yaml_emit(session, &event))
             break;
 
         success = true;
@@ -101,7 +143,7 @@ bool dmi_yaml_mapping_end(dmi_yaml_session_t *session)
     do {
         if (not yaml_mapping_end_event_initialize(&event))
             break;
-        if (not yaml_emitter_emit(session->emitter, &event))
+        if (not dmi_yaml_emit(session, &event))
             break;
 
         success = true;